Double check that test board is reachable!
We can't guarantee that boards are going to be up all the time, so tests
that are effectively integration tests need to be resistant to problems
caused by the board being absent. This adds a check for whether the
board is up to the setup code called in each test that needs an actual
board, skipping if the board can't be pinged (itself a proxy for a
proper test, but a reasonably simple one).