Use a real port in the board test config
Using a bad *remote* port causes weird effects when the board is in a
strange state. Using a real port is better since then it checks how we'd
route that. (Otherwise, when the board is reachable but otherwise
inaccessible, you get bizarre exceptions from the setup phase of the
tests.)