Fix long-standing bug in SpiNNMan with packet reordering
The reading of the P2P table data within GetMachineProcess was wrong, in
that it could reorder the columns of table data. That only occurred when
the reading of one column gets delayed (because of the need to resend a
packet) and the reading of the column after it works immediately,
allowing the results to slip past each other. This would result in them
being entered into the host-side raw P2P table data array in the wrong
order, which in turn would corrupt the building of the parsed P2P table.

Fortunately this is a very rare occurrence in the Python code as the
language is so slow that multiple packets in flight are hardly ever
launched in practice, but it is a bug that is fixed by binding where to
store in the table (known at packet dispatch time!) into the response
callback.