Stop producing quite so many logging messages during testing
Make closed sockets give much less obscure errors too. Complaints about
a negative file descriptor are singularly unhelpful, whereas saying that
a socket is closed is a lot clearer. That in turn helps the connection
listener thread behave more sensibly in the situation, which is what
clears up a lot of error messages.

That this requires looking at the _closed property of sockets is a bit
nasty. We could reproduce that ourselves, but that'd just give an
opportunity for things to go out of synch.