Silence some pylint warnings
The new spalloc client has classes that know plenty about each other. If
I could, I'd lock down the credentials and session key harder, but I
can't because this is Python and I'm not willing to take on the hassle
(authoring, deployment) of writing a compiled extension just for this.